fix: renamed diamond comptroller vip files and folders
