+++ simulations/vip-174/scripts/diamond.ts
