+++ simulations/vip-174/scripts/facet-cut-params-generator.ts
