+++ simulations/vip-174/vip-174-testnet/utils/cut-params.json
