+++ vips/vip-174/vip-174.ts
@@ -1,4 +1,4 @@
-import { cutParams as params } from "../../simulations/vip-diamond-comptroller/vip-diamond-comptroller/utils/cut-params.json";
+import { cutParams as params } from "../../simulations/vip-174/vip-174/utils/cut-params.json";
 import { ProposalType } from "../../src/types";
 import { makeProposal } from "../../src/utils";
 
@@ -35,7 +35,7 @@ const grantAccessControl = () => {
   return accessProposals;
 };
 
-export const vipDiamondComptroller = () => {
+export const vip174 = () => {
   const meta = {
     version: "v1",
     title: "VIP Comptroller Diamond proxy",
